+++ lib/common/common/src/universal_io/mod.rs
@@ -34,8 +34,8 @@ pub use self::simple_disk_cache::{
 pub use self::sorted_block_index::SortedBlockIndex;
 pub use self::traits::{
     CachedReadFs, Item, OpenExtra, OwnedPipeline, ReadPipeline, UniversalAppend, UniversalFlush,
-    UniversalRead, UniversalReadFileOps, UniversalReadFs, UniversalWrite, UniversalWriteFileOps,
-    UserData,
+    UniversalRead, UniversalReadAsync, UniversalReadFileOps, UniversalReadFs, UniversalReadFsAsync,
+    UniversalWrite, UniversalWriteFileOps, UserData,
 };
 pub use self::types::{
     ByteOffset, FileIndex, Flusher, ListedFile, OpenOptions, Populate, ReadBytesItem, ReadRange,
