+++ lib/common/common/src/universal_io/simple_disk_cache/mod.rs
@@ -1,3 +1,4 @@
+mod async_io;
 mod config;
 mod file;
 mod fs;
@@ -14,11 +15,14 @@ pub use file::DiskCache;
 pub use fs::{DiskCacheFs, DiskCacheFsContext};
 
 use crate::mmap::AdviceSetting;
-use crate::universal_io::{OpenOptions, Populate, UniversalRead, UniversalReadFs};
+use crate::universal_io::{OpenOptions, Populate, UniversalReadAsync, UniversalReadFs};
 
 /// Trait bundle for remote backends that can be cached by [`DiskCache`].
+///
+/// Requires [`UniversalReadAsync`]: cache misses and async open-time prefills
+/// fetch from the remote via `read_bytes_async`.
 pub trait DiskCacheRemote:
-    UniversalRead<
+    UniversalReadAsync<
         Fs: Clone + Send + Sync + UniversalReadFs<OpenExtra: Clone + Send + Sync>,
         ReadPipeline<'static, ()>: Send,
         ReadPipeline<'static, Range<u32>>: Send,
@@ -28,7 +32,7 @@ pub trait DiskCacheRemote:
 
 impl<R> DiskCacheRemote for R
 where
-    R: UniversalRead + 'static,
+    R: UniversalReadAsync + 'static,
     R::Fs: Clone + Send + Sync,
     <R::Fs as UniversalReadFs>::OpenExtra: Clone + Send + Sync,
     R::ReadPipeline<'static, ()>: Send,
