+++ lib/segment/src/index/condition_checker.rs
@@ -2,8 +2,6 @@ use common::condition_checker::{
     CheckItem, ConditionChecker, ConstantConditionChecker, Rest, Select, default_check_batched,
 };
 use common::types::PointOffsetType;
-#[cfg(target_os = "linux")]
-use common::universal_io::IoUringFile;
 use common::universal_io::MmapFile;
 
 use crate::common::operation_error::{OperationError, OperationResult};
@@ -55,28 +53,17 @@ pub enum ConditionCheckerEnum<'a> {
     // bool index
     BoolImmutable(BoolCC<'a, ImmutableBoolIndex>),
     BoolMutable(BoolCC<'a, MutableBoolIndex>),
-    #[cfg(target_os = "linux")]
-    BoolRoIoUring(BoolCC<'a, ReadOnlyBoolIndex<IoUringFile>>),
     BoolRoMmap(BoolCC<'a, ReadOnlyBoolIndex<MmapFile>>),
 
     // null index
     NullImmutable(NullCC<'a, ImmutableNullIndex>),
     NullMutable(NullCC<'a, MutableNullIndex>),
-    #[cfg(target_os = "linux")]
-    NullRoIoUring(NullCC<'a, ReadOnlyNullIndex<IoUringFile>>),
     NullRoMmap(NullCC<'a, ReadOnlyNullIndex<MmapFile>>),
 
     // geo index
     GeoRadiusWritable(GeoCC<'a, GeoIndex, GeoRadius>),
     GeoBoundingBoxWritable(GeoCC<'a, GeoIndex, GeoBoundingBox>),
     GeoPolygonWritable(GeoCC<'a, GeoIndex, PolygonWrapper>),
-
-    #[cfg(target_os = "linux")]
-    GeoRadiusRoIoUring(GeoCC<'a, ReadOnlyGeoIndex<IoUringFile>, GeoRadius>),
-    #[cfg(target_os = "linux")]
-    GeoBoundingBoxRoIoUring(GeoCC<'a, ReadOnlyGeoIndex<IoUringFile>, GeoBoundingBox>),
-    #[cfg(target_os = "linux")]
-    GeoPolygonRoIoUring(GeoCC<'a, ReadOnlyGeoIndex<IoUringFile>, PolygonWrapper>),
 
     GeoRadiusRoMmap(GeoCC<'a, ReadOnlyGeoIndex<MmapFile>, GeoRadius>),
     GeoBoundingBoxRoMmap(GeoCC<'a, ReadOnlyGeoIndex<MmapFile>, GeoBoundingBox>),
@@ -87,34 +74,18 @@ pub enum ConditionCheckerEnum<'a> {
     NumericIntWritable(RangeRwCC<'a, IntPayloadType>),
     NumericUuidWritable(RangeRwCC<'a, UuidIntType>),
 
-    #[cfg(target_os = "linux")]
-    NumericFloatRoIoUring(RangeRoCC<'a, IoUringFile, FloatPayloadType>),
-    #[cfg(target_os = "linux")]
-    NumericIntRoIoUring(RangeRoCC<'a, IoUringFile, IntPayloadType>),
-    #[cfg(target_os = "linux")]
-    NumericUuidRoIoUring(RangeRoCC<'a, IoUringFile, UuidIntType>),
-
     NumericFloatRoMmap(RangeRoCC<'a, MmapFile, FloatPayloadType>),
     NumericIntRoMmap(RangeRoCC<'a, MmapFile, IntPayloadType>),
     NumericUuidRoMmap(RangeRoCC<'a, MmapFile, UuidIntType>),
 
     // full-text index
     FullTextWritable(FullTextCC<'a, FullTextIndex>),
-    #[cfg(target_os = "linux")]
-    FullTextRoIoUring(FullTextCC<'a, ReadOnlyFullTextIndex<IoUringFile>>),
     FullTextRoMmap(FullTextCC<'a, ReadOnlyFullTextIndex<MmapFile>>),
 
     // map index
     MapIntWritable(MapWrCC<'a, IntPayloadType>),
     MapStrWritable(MapWrCC<'a, str>),
     MapUuidWritable(MapWrCC<'a, UuidIntType>),
-
-    #[cfg(target_os = "linux")]
-    MapIntRoIoUring(MapRoCC<'a, IoUringFile, IntPayloadType>),
-    #[cfg(target_os = "linux")]
-    MapStrRoIoUring(MapRoCC<'a, IoUringFile, str>),
-    #[cfg(target_os = "linux")]
-    MapUuidRoIoUring(MapRoCC<'a, IoUringFile, UuidIntType>),
 
     MapIntRoMmap(MapRoCC<'a, MmapFile, IntPayloadType>),
     MapStrRoMmap(MapRoCC<'a, MmapFile, str>),
@@ -142,51 +113,27 @@ impl ConditionChecker for ConditionCheckerEnum<'_> {
             Self::TestBitOfId(c) => c.check_batched(ids, select, rest),
             Self::BoolImmutable(c) => c.check_batched(ids, select, rest),
             Self::BoolMutable(c) => c.check_batched(ids, select, rest),
-            #[cfg(target_os = "linux")]
-            Self::BoolRoIoUring(c) => c.check_batched(ids, select, rest),
             Self::BoolRoMmap(c) => c.check_batched(ids, select, rest),
             Self::NullImmutable(c) => c.check_batched(ids, select, rest),
             Self::NullMutable(c) => c.check_batched(ids, select, rest),
-            #[cfg(target_os = "linux")]
-            Self::NullRoIoUring(c) => c.check_batched(ids, select, rest),
             Self::NullRoMmap(c) => c.check_batched(ids, select, rest),
             Self::GeoRadiusWritable(c) => c.check_batched(ids, select, rest),
             Self::GeoBoundingBoxWritable(c) => c.check_batched(ids, select, rest),
             Self::GeoPolygonWritable(c) => c.check_batched(ids, select, rest),
-            #[cfg(target_os = "linux")]
-            Self::GeoRadiusRoIoUring(c) => c.check_batched(ids, select, rest),
-            #[cfg(target_os = "linux")]
-            Self::GeoBoundingBoxRoIoUring(c) => c.check_batched(ids, select, rest),
-            #[cfg(target_os = "linux")]
-            Self::GeoPolygonRoIoUring(c) => c.check_batched(ids, select, rest),
             Self::GeoRadiusRoMmap(c) => c.check_batched(ids, select, rest),
             Self::GeoBoundingBoxRoMmap(c) => c.check_batched(ids, select, rest),
             Self::GeoPolygonRoMmap(c) => c.check_batched(ids, select, rest),
             Self::NumericFloatWritable(c) => c.check_batched(ids, select, rest),
             Self::NumericIntWritable(c) => c.check_batched(ids, select, rest),
             Self::NumericUuidWritable(c) => c.check_batched(ids, select, rest),
-            #[cfg(target_os = "linux")]
-            Self::NumericFloatRoIoUring(c) => c.check_batched(ids, select, rest),
-            #[cfg(target_os = "linux")]
-            Self::NumericIntRoIoUring(c) => c.check_batched(ids, select, rest),
-            #[cfg(target_os = "linux")]
-            Self::NumericUuidRoIoUring(c) => c.check_batched(ids, select, rest),
             Self::NumericFloatRoMmap(c) => c.check_batched(ids, select, rest),
             Self::NumericIntRoMmap(c) => c.check_batched(ids, select, rest),
             Self::NumericUuidRoMmap(c) => c.check_batched(ids, select, rest),
             Self::FullTextWritable(c) => c.check_batched(ids, select, rest),
-            #[cfg(target_os = "linux")]
-            Self::FullTextRoIoUring(c) => c.check_batched(ids, select, rest),
             Self::FullTextRoMmap(c) => c.check_batched(ids, select, rest),
             Self::MapIntWritable(c) => c.check_batched(ids, select, rest),
             Self::MapStrWritable(c) => c.check_batched(ids, select, rest),
             Self::MapUuidWritable(c) => c.check_batched(ids, select, rest),
-            #[cfg(target_os = "linux")]
-            Self::MapIntRoIoUring(c) => c.check_batched(ids, select, rest),
-            #[cfg(target_os = "linux")]
-            Self::MapStrRoIoUring(c) => c.check_batched(ids, select, rest),
-            #[cfg(target_os = "linux")]
-            Self::MapUuidRoIoUring(c) => c.check_batched(ids, select, rest),
             Self::MapIntRoMmap(c) => c.check_batched(ids, select, rest),
             Self::MapStrRoMmap(c) => c.check_batched(ids, select, rest),
             Self::MapUuidRoMmap(c) => c.check_batched(ids, select, rest),
@@ -206,51 +153,27 @@ impl ConditionChecker for ConditionCheckerEnum<'_> {
             Self::TestBitOfId(c) => c.check(point_id),
             Self::BoolImmutable(c) => c.check(point_id),
             Self::BoolMutable(c) => c.check(point_id),
-            #[cfg(target_os = "linux")]
-            Self::BoolRoIoUring(c) => c.check(point_id),
             Self::BoolRoMmap(c) => c.check(point_id),
             Self::NullImmutable(c) => c.check(point_id),
             Self::NullMutable(c) => c.check(point_id),
-            #[cfg(target_os = "linux")]
-            Self::NullRoIoUring(c) => c.check(point_id),
             Self::NullRoMmap(c) => c.check(point_id),
             Self::GeoRadiusWritable(c) => c.check(point_id),
             Self::GeoBoundingBoxWritable(c) => c.check(point_id),
             Self::GeoPolygonWritable(c) => c.check(point_id),
-            #[cfg(target_os = "linux")]
-            Self::GeoRadiusRoIoUring(c) => c.check(point_id),
-            #[cfg(target_os = "linux")]
-            Self::GeoBoundingBoxRoIoUring(c) => c.check(point_id),
-            #[cfg(target_os = "linux")]
-            Self::GeoPolygonRoIoUring(c) => c.check(point_id),
             Self::GeoRadiusRoMmap(c) => c.check(point_id),
             Self::GeoBoundingBoxRoMmap(c) => c.check(point_id),
             Self::GeoPolygonRoMmap(c) => c.check(point_id),
             Self::NumericFloatWritable(c) => c.check(point_id),
             Self::NumericIntWritable(c) => c.check(point_id),
             Self::NumericUuidWritable(c) => c.check(point_id),
-            #[cfg(target_os = "linux")]
-            Self::NumericFloatRoIoUring(c) => c.check(point_id),
-            #[cfg(target_os = "linux")]
-            Self::NumericIntRoIoUring(c) => c.check(point_id),
-            #[cfg(target_os = "linux")]
-            Self::NumericUuidRoIoUring(c) => c.check(point_id),
             Self::NumericFloatRoMmap(c) => c.check(point_id),
             Self::NumericIntRoMmap(c) => c.check(point_id),
             Self::NumericUuidRoMmap(c) => c.check(point_id),
             Self::FullTextWritable(c) => c.check(point_id),
-            #[cfg(target_os = "linux")]
-            Self::FullTextRoIoUring(c) => c.check(point_id),
             Self::FullTextRoMmap(c) => c.check(point_id),
             Self::MapIntWritable(c) => c.check(point_id),
             Self::MapStrWritable(c) => c.check(point_id),
             Self::MapUuidWritable(c) => c.check(point_id),
-            #[cfg(target_os = "linux")]
-            Self::MapIntRoIoUring(c) => c.check(point_id),
-            #[cfg(target_os = "linux")]
-            Self::MapStrRoIoUring(c) => c.check(point_id),
-            #[cfg(target_os = "linux")]
-            Self::MapUuidRoIoUring(c) => c.check(point_id),
             Self::MapIntRoMmap(c) => c.check(point_id),
             Self::MapStrRoMmap(c) => c.check(point_id),
             Self::MapUuidRoMmap(c) => c.check(point_id),
@@ -270,51 +193,27 @@ impl ConditionChecker for ConditionCheckerEnum<'_> {
             Self::TestBitOfId(c) => c.check_infallible(point_id),
             Self::BoolImmutable(c) => c.check_infallible(point_id),
             Self::BoolMutable(c) => c.check_infallible(point_id),
-            #[cfg(target_os = "linux")]
-            Self::BoolRoIoUring(c) => c.check_infallible(point_id),
             Self::BoolRoMmap(c) => c.check_infallible(point_id),
             Self::NullImmutable(c) => c.check_infallible(point_id),
             Self::NullMutable(c) => c.check_infallible(point_id),
-            #[cfg(target_os = "linux")]
-            Self::NullRoIoUring(c) => c.check_infallible(point_id),
             Self::NullRoMmap(c) => c.check_infallible(point_id),
             Self::GeoRadiusWritable(c) => c.check_infallible(point_id),
             Self::GeoBoundingBoxWritable(c) => c.check_infallible(point_id),
             Self::GeoPolygonWritable(c) => c.check_infallible(point_id),
-            #[cfg(target_os = "linux")]
-            Self::GeoRadiusRoIoUring(c) => c.check_infallible(point_id),
-            #[cfg(target_os = "linux")]
-            Self::GeoBoundingBoxRoIoUring(c) => c.check_infallible(point_id),
-            #[cfg(target_os = "linux")]
-            Self::GeoPolygonRoIoUring(c) => c.check_infallible(point_id),
             Self::GeoRadiusRoMmap(c) => c.check_infallible(point_id),
             Self::GeoBoundingBoxRoMmap(c) => c.check_infallible(point_id),
             Self::GeoPolygonRoMmap(c) => c.check_infallible(point_id),
             Self::NumericFloatWritable(c) => c.check_infallible(point_id),
             Self::NumericIntWritable(c) => c.check_infallible(point_id),
             Self::NumericUuidWritable(c) => c.check_infallible(point_id),
-            #[cfg(target_os = "linux")]
-            Self::NumericFloatRoIoUring(c) => c.check_infallible(point_id),
-            #[cfg(target_os = "linux")]
-            Self::NumericIntRoIoUring(c) => c.check_infallible(point_id),
-            #[cfg(target_os = "linux")]
-            Self::NumericUuidRoIoUring(c) => c.check_infallible(point_id),
             Self::NumericFloatRoMmap(c) => c.check_infallible(point_id),
             Self::NumericIntRoMmap(c) => c.check_infallible(point_id),
             Self::NumericUuidRoMmap(c) => c.check_infallible(point_id),
             Self::FullTextWritable(c) => c.check_infallible(point_id),
-            #[cfg(target_os = "linux")]
-            Self::FullTextRoIoUring(c) => c.check_infallible(point_id),
             Self::FullTextRoMmap(c) => c.check_infallible(point_id),
             Self::MapIntWritable(c) => c.check_infallible(point_id),
             Self::MapStrWritable(c) => c.check_infallible(point_id),
             Self::MapUuidWritable(c) => c.check_infallible(point_id),
-            #[cfg(target_os = "linux")]
-            Self::MapIntRoIoUring(c) => c.check_infallible(point_id),
-            #[cfg(target_os = "linux")]
-            Self::MapStrRoIoUring(c) => c.check_infallible(point_id),
-            #[cfg(target_os = "linux")]
-            Self::MapUuidRoIoUring(c) => c.check_infallible(point_id),
             Self::MapIntRoMmap(c) => c.check_infallible(point_id),
             Self::MapStrRoMmap(c) => c.check_infallible(point_id),
             Self::MapUuidRoMmap(c) => c.check_infallible(point_id),
