+++ lib/segment/src/index/universal_io.rs
@@ -1,8 +1,8 @@
-#[cfg(target_os = "linux")]
-use common::universal_io::IoUringFile;
 #[cfg(test)]
 use common::universal_io::ReadOnly;
-use common::universal_io::{DiskCache, DiskCacheRemote, MmapFile, UniversalRead};
+use common::universal_io::{
+    DiskCache, DiskCacheRemote, MmapFile, UniversalRead, UniversalReadFsAsync,
+};
 
 use crate::index::condition_checker::ConditionCheckerEnum;
 use crate::index::field_index::bool_index::{BoolConditionChecker, ReadOnlyBoolIndex};
@@ -29,8 +29,16 @@ type RangeCC<'a, S, T> = RangeConditionChecker<'a, ReadOnlyNumericIndexInner<T, 
 /// Sometimes generic-over-[`UniversalRead`] code must branch on the concrete
 /// UIO type. This trait is a place for such branching. Any implementations of
 /// [`UniversalRead`] used in this [crate] should implement this.
+///
+/// The `Fs: UniversalReadFsAsync` bound makes every read-only-segment backend
+/// async-capable at the filesystem level — [`CachedFs`] (which every segment
+/// open and live-reload goes through) parks `open_async` futures in its
+/// prefetch pool. Notably io_uring implements no async surface and is
+/// therefore not a read-only-segment backend.
+///
+/// [`CachedFs`]: common::universal_io::CachedFs
 #[rustfmt::skip]
-pub trait UniversalReadExt: UniversalRead {
+pub trait UniversalReadExt: UniversalRead<Fs: UniversalReadFsAsync> {
     // Methods to convert generic `BlahBlahConditionChecker<S>` -> non-generic enum.
     fn condition_checker_bool            <'a>(i: BoolCC<'a, Self>)                    -> EnumCC<'a>;
     fn condition_checker_full_text       <'a>(i: FullTextCC<'a, Self>)                -> EnumCC<'a>;
@@ -62,23 +70,6 @@ impl UniversalReadExt for MmapFile {
     fn condition_checker_numeric_float   <'a>(i: RangeCC<'a, Self, FloatPayloadType>) -> EnumCC<'a> { EnumCC::NumericFloatRoMmap  (i) }
     fn condition_checker_numeric_int     <'a>(i: RangeCC<'a, Self, IntPayloadType>)   -> EnumCC<'a> { EnumCC::NumericIntRoMmap    (i) }
     fn condition_checker_numeric_uuid    <'a>(i: RangeCC<'a, Self, UuidIntType>)      -> EnumCC<'a> { EnumCC::NumericUuidRoMmap   (i) }
-}
-
-#[cfg(target_os = "linux")]
-#[rustfmt::skip]
-impl UniversalReadExt for IoUringFile {
-    fn condition_checker_bool            <'a>(i: BoolCC<'a, Self>)                    -> EnumCC<'a> { EnumCC::BoolRoIoUring          (i) }
-    fn condition_checker_full_text       <'a>(i: FullTextCC<'a, Self>)                -> EnumCC<'a> { EnumCC::FullTextRoIoUring      (i) }
-    fn condition_checker_geo_bounding_box<'a>(i: GeoCC<'a, Self, GeoBoundingBox>)     -> EnumCC<'a> { EnumCC::GeoBoundingBoxRoIoUring(i) }
-    fn condition_checker_geo_polygon     <'a>(i: GeoCC<'a, Self, PolygonWrapper>)     -> EnumCC<'a> { EnumCC::GeoPolygonRoIoUring    (i) }
-    fn condition_checker_geo_radius      <'a>(i: GeoCC<'a, Self, GeoRadius>)          -> EnumCC<'a> { EnumCC::GeoRadiusRoIoUring     (i) }
-    fn condition_checker_map_int         <'a>(i: MapCC<'a, Self, IntPayloadType>)     -> EnumCC<'a> { EnumCC::MapIntRoIoUring        (i) }
-    fn condition_checker_map_str         <'a>(i: MapCC<'a, Self, str>)                -> EnumCC<'a> { EnumCC::MapStrRoIoUring        (i) }
-    fn condition_checker_map_uuid        <'a>(i: MapCC<'a, Self, UuidIntType>)        -> EnumCC<'a> { EnumCC::MapUuidRoIoUring       (i) }
-    fn condition_checker_null            <'a>(i: NullCC<'a, Self>)                    -> EnumCC<'a> { EnumCC::NullRoIoUring          (i) }
-    fn condition_checker_numeric_float   <'a>(i: RangeCC<'a, Self, FloatPayloadType>) -> EnumCC<'a> { EnumCC::NumericFloatRoIoUring  (i) }
-    fn condition_checker_numeric_int     <'a>(i: RangeCC<'a, Self, IntPayloadType>)   -> EnumCC<'a> { EnumCC::NumericIntRoIoUring    (i) }
-    fn condition_checker_numeric_uuid    <'a>(i: RangeCC<'a, Self, UuidIntType>)      -> EnumCC<'a> { EnumCC::NumericUuidRoIoUring   (i) }
 }
 
 // TODO: add enum variants for these
